Fixing state machine and cache, and starting frontend

+++ laravel-backend/app/Services/BaseService.php
@@ -5,78 +5,86 @@ namespace App\Services;
 use App\Http\Requests\SearchObjects\BaseSearchObject;
 use App\Services\Contracts\BaseServiceInterface;
 use Exception;
+use Illuminate\Support\Facades\Validator;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Query\Builder;
+use Illuminate\Http\Request;
 
 abstract class BaseService implements BaseServiceInterface
 {
     abstract protected function getModelClass();
+    abstract function getInsertRequestClass();
+    abstract function getUpdateRequestClass();
 
-    protected function handleDeleteResponse()
+
+    public function getPageable($searchObject)
     {
-        return response()->noContent();
+        $query = $this->getModelClass()->query();
+
+        $query = $this->includeRelation($searchObject, $query);
+        $query = $this->addFilter($searchObject, $query);
+
+        return $query->paginate($searchObject->size);
     }
 
-    public function getAll()
+
+    public function getById(int $id, $searchObject)
     {
-        $searchObjectInstance = app($this->getSearchObject());
-        $searchObjectInstance->fill(request()->query());
-
-        $query = $this->getModelInstance()->query();
-
-        $query = $this->includeRelation($searchObjectInstance, $query);
-        $query = $this->addFilter($searchObjectInstance, $query);
-
-        return $query->paginate($searchObjectInstance->size);
-    }
-
-    public function getById(int $id)
-    {
-        $searchObjectInstance = app($this->getSearchObject());
-        $query = $this->getModelInstance()->query();
-
-        $searchObjectInstance->fill(request()->query());
-        $query = $this->includeRelation($searchObjectInstance, $query);
-        $query = $this->addFilter($searchObjectInstance, $query);
-
+        $query = $this->getModelClass()->query();
+        $query = $this->includeRelation($searchObject, $query);
         $result = $query->find($id);
 
         if (!$result) {
-            throw new Exception("Resource not found!");
+            throw new Exception("Error!");
         }
 
         return $result;
     }
 
-    public function add(array $request)
+    public function add(Request $request)
     {
-        return $this->getModelInstance()->create($request);
+        $this->validateRequest($request, $this->getInsertRequestClass());
+        return $this->getModelInstance()->create($request->all());
     }
 
-    public function update(array $request, int $id)
+    public function update(Request $request, int $id)
     {
-        $model = $this->getModelInstance()->findOrFail($id);
-        $model->update($request);
+        $this->validateRequest($request, $this->getUpdateRequestClass());
+        $model = $this->getModelInstance()->find($id);
+
+        if (!$model) {
+            throw new Exception("Error!");
+        }
+
+        $model->update($request->all());
+
         return $model;
     }
 
     public function remove(int $id)
     {
-        $model = $this->getModelInstance()->findOrFail($id);
+        $model = $this->getModelInstance()->find($id);
+
+        if (!$model) {
+            throw new Exception("Error!");
+        }
+
         $model->delete();
-        return $this->handleDeleteResponse();
+
+        return $model;
     }
 
-    public function addFilter($searchObject, $query)
+    public function addFilter($searchObject, Builder $query)
     {
         return $query;
     }
 
-    public function getSearchObject()
+    public function getSearchObject($params)
     {
-        return BaseSearchObject::class;
+        return new BaseSearchObject($params);
     }
 
-    public function includeRelation($searchObject, $query)
+    public function includeRelation($searchObject, Builder $query)
     {
         return $query;
     }
@@ -84,6 +92,19 @@ abstract class BaseService implements BaseServiceInterface
     protected function getModelInstance(): Model
     {
         $modelClass = $this->getModelClass();
+
         return new $modelClass;
+    }
+
+    public function validateRequest(Request $request, $formRequest)
+    {
+        $formRequestInstance = new $formRequest();
+        $rules = $formRequestInstance->rules();
+        $validator = Validator::make($request->all(), $rules);
+
+        if ($validator->fails()) {
+            throw new Exception("Error");
+        }
+        return $validator->validated();
     }
 }
